script minor improvement

+++ bootstrap.sh
@@ -22,10 +22,13 @@ ripgrep \
 htop
 
 # install broot
-curl -o broot -L https://dystroy.org/broot/download/x86_64-linux/broot
-sudo mv broot /usr/local/bin
-sudo chmod +x /usr/local/bin/broot
-broot
+if [ ! -f "/usr/local/bin/broot" ]
+then
+    curl -o broot -L https://dystroy.org/broot/download/x86_64-linux/broot
+    sudo mv broot /usr/local/bin
+    sudo chmod +x /usr/local/bin/broot
+    broot
+fi
 
 ## dpkg -l tree htop curl git zsh vim tmux
 # curl -V, git version, zsh --version, vim --version
